+++ OUDS/Core/Components/Sources/Actions/Button/Internal/ButtonStyle.swift
@@ -19,28 +19,26 @@ import SwiftUI
 ///
 /// Four appearances are proposed:
 /// - **default**: Default buttons are used for actions which are not mandatory or essential for the user.
-///
 /// - **strong**: The strong "call for action" on the page should be singular and prominent, limited to one per view.
 /// It should be reserved for the most critical action, such as "Next," "Save," "Submit," etc.
-///
 /// - **minimal**: Minimal buttons are commonly used for actions that are considered less crucial. They can be used independently or together with a strong button.
-///
 /// - **negative**: Negative buttons should be used sparingly to warn of a destructive action, for example, delete or remove, typically
 /// resulting in the opening of a confirmation dialog.
 ///
 /// Two styles are available:
 /// - **normal**: is the normal state of a button which can be disabled, pressed, hovered or normal (i.e. enabled)
 /// - **loading**: the style used after button was clicked and probably data are requested before navigate to a next screen or get updated data.
-///
 struct StyleForButton: ButtonStyle {
 
-    // MARK: Stored Properties
+    // MARK: Properties
 
     private let appearance: OUDSButton.Appearance
     private let style: OUDSButton.Style
+    private let size: OUDSButton.Size
     private let isHover: Bool
     private let isFullWidth: Bool
 
+    @Environment(\.theme) private var theme
     @Environment(\.isEnabled) private var isEnabled
 
     // MARK: Initializer
@@ -51,11 +49,18 @@ struct StyleForButton: ButtonStyle {
     /// - Parameters:
     ///    - appearance: The button appearance
     ///    - style: The button style
+    ///    - size: The size of the button
     ///    - isHover: Flag if button is hovered (e.g. by mouse)
     ///    - isFullWidth: Flag to rise to make the button take all screen width
-    init(appearance: OUDSButton.Appearance, style: OUDSButton.Style, isHover: Bool, isFullWidth: Bool) {
+    init(appearance: OUDSButton.Appearance,
+         style: OUDSButton.Style,
+         size: OUDSButton.Size,
+         isHover: Bool,
+         isFullWidth: Bool)
+    {
         self.appearance = appearance
         self.style = style
+        self.size = size
         self.isHover = isHover
         self.isFullWidth = isFullWidth
     }
@@ -67,15 +72,40 @@ struct StyleForButton: ButtonStyle {
         case .default:
             configuration
                 .label
-                .frame(maxWidth: isFullWidth ? .infinity : nil)
+                .frame(minWidth: minWidth,
+                       maxWidth: maxWidth,
+                       minHeight: minHeight,
+                       alignment: .center)
                 .contentShape(Rectangle())
                 .modifier(ButtonViewModifier(appearance: appearance, state: internalState(isPressed: configuration.isPressed)))
         case .loading:
             configuration.label
-                .frame(maxWidth: isFullWidth ? .infinity : nil)
+                .frame(maxWidth: maxWidth)
                 .modifier(ButtonViewModifier(appearance: appearance, state: .loading))
-                .modifier(ButtonLoadingContentModifier(appearance: appearance))
+                .modifier(ButtonLoadingContentModifier(appearance: appearance, size: size))
         }
+    }
+
+    private var minWidth: CGFloat {
+        switch size {
+        case .default:
+            theme.button.sizeMinWidthDefault
+        case .small:
+            theme.button.sizeMinWidthSmall
+        }
+    }
+
+    private var minHeight: CGFloat {
+        switch size {
+        case .default:
+            theme.button.sizeMinHeightDefault
+        case .small:
+            theme.button.sizeMinHeightSmall
+        }
+    }
+
+    private var maxWidth: CGFloat? {
+        isFullWidth ? .infinity : nil
     }
 
     private func internalState(isPressed: Bool) -> ButtonInternalState {