+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `button` component to version 3.3.0 (Orange-OpenSource/ouds-ios#1583)
 - **BREAKING**: `link` component to v2.4.0 (Orange-OpenSource/ouds-ios#1640)
 - `suggestion chip` and `filter chip` components to v1.5.0 (Orange-OpenSource/ouds-ios#1588)
 - **BREAKING**: `link` component to v2.3.0 (Orange-OpenSource/ouds-ios#1589)